Moved code around for clarity
