+++ skate-plugin/src/main/kotlin/com/slack/sgp/intellij/codeowners/model/Path.kt
@@ -1,0 +1,6 @@
+package com.slack.sgp.intellij.codeowners.model
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class Path(val path: String, val notify: Boolean?)
